feat: mathjax 归入插件机制，可选择使用
