+++ src/utils/pluginCenter.js
@@ -1,0 +1,4 @@
+// 检测插件是否安装
+export default {
+  mathjax: false,
+};
